requirements.txt
aioredis==1.3.1
appnope==0.1.2
asgiref==3.5.0
astroid==2.9.3
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backcall==0.2.0
bcrypt==3.2.0
black==22.1.0
bleach==4.1.0
boto3==1.20.49
botocore==1.23.49
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.11
click==8.0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
daphne==3.0.2
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.12
django-appconf==1.0.5
django-braces==1.15.0
django-countries==7.2.1
django-extensions==3.1.5
django-filter==21.1
django-imagekit==4.1.0
django-oauth-toolkit==1.7.0
django-rest-framework-social-oauth2==1.1.0
django-storages==1.12.3
django-summernote==0.8.20.0
djangorestframework==3.13.1
drf-nested-routers==0.93.4
drf-yasg==1.20.0
executing==0.8.2
filelock==3.4.2
hiredis==2.0.0
hyperlink==21.0.0
identify==2.4.9
idna==3.3
incremental==21.3.0
inflection==0.5.1
ipdb==0.13.9
ipython==8.0.1
isort==5.10.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jwcrypto==1.0
lazy-object-proxy==1.7.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
msgpack==1.0.3
mypy-extensions==0.4.3
mysqlclient==2.1.0
nodeenv==1.6.0
oauthlib==3.2.0
packaging==21.3
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.0.1
platformdirs==2.4.1
pre-commit==2.17.0
prompt-toolkit==3.0.27
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyright==0.0.13
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
s3transfer==0.5.1
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
stack-data==0.1.4
toml==0.10.2
tomli==2.0.0
traitlets==5.1.1
Twisted==22.1.0
txaio==21.2.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
virtualenv==20.13.1
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.13.3
zope.interface==5.4.0
